Move debug info to sub-module within `fj-interop`
