+++ fj-operations/src/lib.rs
@@ -13,7 +13,7 @@ mod sketch;
 mod sweep;
 mod transform;
 
-use fj_interop::DebugInfo;
+use fj_interop::debug::DebugInfo;
 use fj_kernel::shape::Shape;
 use fj_math::{Aabb, Scalar};
