+++ fj-operations/src/sweep.rs
@@ -1,4 +1,4 @@
-use fj_interop::DebugInfo;
+use fj_interop::debug::DebugInfo;
 use fj_kernel::{algorithms::sweep_shape, shape::Shape};
 use fj_math::{Aabb, Scalar, Vector};
