+++ fj-operations/src/transform.rs
@@ -1,4 +1,4 @@
-use fj_interop::DebugInfo;
+use fj_interop::debug::DebugInfo;
 use fj_kernel::shape::Shape;
 use fj_math::{Aabb, Scalar, Transform};
 use parry3d_f64::math::Isometry;
